requirements.txt
aiogram==3.*
appwrite==7.*
python-dotenv
pydantic==2.*
gspread
google-auth
